修改log bug

+++ InvoicingSystemClient/src/po/MemberPO.java
@@ -13,14 +13,14 @@ public class MemberPO implements Serializable{
 	int category;
 	long defaultOperatorID;
 	String name;
-	int phoneNum;
+	long phoneNum;
 	String address;
 	String eMail;
 	double shouldPay;
 	double shouldGet;
 	double quota;//”¶ ’∂Ó∂»
 	
-	public MemberPO(long id,int rank,int postCode,int category,String name,int phoneNum,
+	public MemberPO(long id,int rank,int postCode,int category,String name,long phoneNum,
 	String address,String eMail,double shouldPay,double shouldGet,double quota,long defaultOperatorID){
 		this.id=id;
 		this.rank=rank;
@@ -50,7 +50,7 @@ public class MemberPO implements Serializable{
 	public String getName() {
 		return name;
 	}
-	public int getPhoneNum(){
+	public long getPhoneNum(){
 		return phoneNum;
 	}
 	public String getAddress(){
